Vibe coding went okay, but not perfect
So this refactor process worked very well for being functionality, but the file format by default wasn't the way I'd maintain data (most with the file size of decisionHelperSlice and it relying on comments). However, I'm not sure if AI prefers content all located in 1 file, and most edits I plan to ask for should be simple requests. But if AI struggles to work with this large file for most requests, I'll refactor it in 4 smaller files for each table to make it easier to separate duties and work with (for my eyes and without AI).