Vibe Coding a Design Emailer
Made something for making email templates. Had to do some extra asks so the HTML used <table> tags instead of <style> ones to highlight information. Another edit will require there to be specific selecting and editing of pre-existing designs, because a problem is that the formatting that looks ok on the web ends up not being ok for web/mobile (ex: whitespace looks worse on mail than webpages)