Vibe coding image analyzer
So Image analyzer right now is bad. But the issue itself isn't the code quality, but the code approach. LLMs themselves just don't seem to be good enough to do the job of talking about an image. So I'm going to work on setting up Kubernetes to set up and launch 3 servers now (I'm doing the website and file server individually in terminal). This will help launch all the apps at once