Mockup of google trends marketing idea
Definitely needs a lot of work. Specifically with the "what should this direct users to do". But I think one clear direction that will happen (and need at least some technical work) would be storing and proposing the specific affiliate links we'd like to be recommended as the output from these tools. Also, with all of these edits in general, could definitely get code optimizations and refactoring to be easier for others to edit, but at this point not as important as just growing fast which it can rn (still a small enough feature, so context isn't a problem yet for cursor tool)